Extract IPC dispatch from main.rs into ipc_dispatch.rs
Move process_ipc_commands(), dispatch_ipc(), resolve_surface_mut(),
and resolve_surface() (~660 lines) into a dedicated module. This
continues the modularization of amux-app (issue #42).
